Click functionality on co-pi network and bar graph nodes
The user is not able to view highlighted entries in the bar graph which
are outside the screen when a node in the force network is highlighted.
A click functionality allows all the relevant entries to remain
highlighted for ease of use of the user until the node is clicked again
or the background of the visualization is clicked, meaning - first click
on nodes highlights , second click on the same node clears or click on
background clears.